requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
demucs==4.0.1
docopt==0.6.2
dora-search==0.1.12
einops==0.7.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
huggingface-hub==0.19.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lameenc==1.6.3
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.2
lightning-utilities==0.10.0
llvmlite==0.41.1
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mutagen==1.47.0
networkx==3.2.1
noisereduce==3.0.0
numba==0.58.1
numpy==1.26.2
omegaconf==2.3.0
onnxruntime==1.16.3
openunmix==1.2.1
optuna==3.4.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
primePy==1.3
protobuf==4.25.1
pyannote.audio==3.1.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.2
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.4
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.23
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
torch==2.1.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1
tqdm==4.66.1
treetable==0.2.5
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websockets==12.0
yarl==1.9.3
yt-dlp==2023.11.16
